requirements.txt
cs50
Flask
Flask-Session
requests
tempfile
werkzeug.exceptions
werkzeug.security